requirements.txt
openai
gspread
oauth2client
python-dotenv
requests
google-auth
